Tidying up
git-svn-id: svn+ssh://svn.berlios.de/svnroot/repos/minos/trunk@218 5aba93d0-1d48-0410-ab36-a573633a0585

+++ Minos/MinosLogger/SettingsEditor.cpp
@@ -216,7 +216,7 @@ void __fastcall TSettingsEditDlg::GridHintTimerTimer( TObject */*Sender*/ )
          if ( PtInRect( &( QGrid->ClientRect ), mpos2 ) )
          {
             std::vector<std::string>hints = bundle->getBundleHints( );
-            if (HintRow < hints.size())
+            if (HintRow < (int)hints.size())
             {
                QGrid->Hint = hints[ HintRow - 1 ].c_str();
                GridHintWindow->SetXY( mpos.x, mpos.y );
